Refactor Snake class and add raw mode support
So, previously the training would only run for 188-195 episodes because there was a bug in the reset function. The reset function didn't cleared the apples of previous episode. This means that, it might happen that the snake's body is on a apple. And the spawnFood function checked 2 conditions, the generated apple index should not be occupied by snake's body 'and' that index must be 0 in the grid. The second condition was causing the bug also, so I removed it and rewrote it. I also simplified the move() function.